LayoutTests/platform/mac-catalina-wk2/imported/w3c/web-platform-tests/websockets/Send-65K-data.any-expected.txt


FAIL Send 65K data on a WebSocket - Connection should be closed assert_equals: expected 0 but got 65000
